+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add `print_selected_statements()` presenter for inspecting `PolisClusteringResult`.
 - Add `print_consensus_statements()` presenter for inspecting `PolisClusteringResult`.
 - Allow `pick_max` and `confidence` interval args to be set in `polis.run_clustering()`.
+- Allow `get_corrected_centroid_guesses()` to unflip each axis if correction not needed.
 
 ### Chores